Extend the supporting superclass to support upgrade to rails 6.1
I fount that the Rails 6.1 changed some superclass of active_storage
and action_text.
please refer to this rails commit:
https://github.com/rails/rails/commit/a50d2f1993a3197d98535e6c9fceb5fedab78ee9

And it will make the rails app boot failure when the migrant want to do
the schema process.

I add a extend version to make it support for rails 6.1 above app, that
it can refer to the correct superclass.
If the project still under Rails 6.0, then it should only use the
original version as the rails library doesn't have that two classes:
ActiveStorge::Record and ActionText::Record.